edycja oferty, oferty użytkownika, dodawanie oferty, widok oferty, edycja profilu

+++ src/AppBundle/Controller/JobController.php
@@ -45,12 +45,12 @@ class JobController extends Controller
                 ->createQueryBuilder('j')
                 ->SELECT('j')
                 ->where('j.position LIKE :query')
-                ->orWhere('j.description LIKE :query')
+                ->orWhere('j.companyName LIKE :query')
                 ->orWhere('j.location LIKE :query')
+                ->andWhere('j.verified = true')
                 ->setParameter('query', '%' . $query . '%')
                 ->getQuery()
                 ->getResult();
-
 
 
         return $this->render('Job/search.html.twig', [
@@ -65,8 +65,11 @@ class JobController extends Controller
     public function deleteAction(Job $job)
     {
 
-        $em = $this->getDoctrine()
-                ->getManager();
+        if (!$job) {
+            throw $this->createNotFoundException('Oferta nie istnieje');
+        }
+
+        $em = $this->getDoctrine()->getManager();
 
         $em->remove($job);
         $em->persist();
@@ -79,9 +82,56 @@ class JobController extends Controller
      */
     public function addAction(Request $request)
     {
+        $user = $this->getUser();
+
+        $userNbJobs = $this->getDoctrine()
+                ->getRepository('AppBundle:Job')
+                ->findBy(array('user' => $user));
+
         $job = new Job();
+        $job->setEmail($user->getEmail());
+        $job->setUser($user);
+
         $form = $this->createForm(new JobType(), $job);
         $form->handleRequest($request);
+
+        if ($form->isSubmitted() && !$user) {
+            $this->addFlash('error', 'Aby dodać ofertę musisz być zalogowany');
+            return $this->redirectToRoute('jobs_list');
+        }
+        if ($form->isValid()) {
+            if (count($userNbJobs) > 2 && $user->getVerified() == false) {
+
+                $user->setVerified(true);
+                $job->setVerified(true);
+                $this->addFlash('notice', 'Oferta została pomyślnie dodana oraz użytkownik został'
+                        . ' automatycznie zweryfikowany jako zaufany');
+
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($job);
+                $em->flush();
+
+                return $this->redirectToRoute('jobs_list');
+            }
+            if ($user->getVerified() == true) {
+
+                $job->setVerified(true);
+                $this->addFlash('notice', 'Oferta została pomyślnie dodana');
+
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($job);
+                $em->flush();
+
+                return $this->redirectToRoute('jobs_list');
+            }
+
+            $em = $this->getDoctrine()->getManager();
+            $em->persist($job);
+            $em->flush();
+
+            $this->addFlash('notice', 'Oferta została pomyślnie dodana i oczekuje na moderacje');
+            return $this->redirectToRoute('jobs_list');
+        }
 
         return $this->render('Job/add.html.twig', array(
                     'form' => $form->createView(),
@@ -89,25 +139,95 @@ class JobController extends Controller
     }
 
     /**
-     * @Route("/edytuj/{id}")
-     * @Template()
+     * @Route("/edytuj/{id}", name="edit_job")
+     * 
      */
-    public function editAction($id)
+    public function editAction(Request $request, Job $job)
     {
-        return array(
-                // ...
-        );
+        $user = $this->getUser();
+
+        $form = $this->createForm(new JobType(), $job);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && !$user) {
+            $this->addFlash('notice', 'Musisz być zalogowany aby edytować ofertę');
+            return $this->redirectToRoute('jobs_list');
+        }
+
+        $userNbJobs = $this->getDoctrine()
+                ->getRepository('AppBundle:Job')
+                ->findBy(array('user' => $user));
+
+        if ($form->isValid()) {
+            if (count($userNbJobs) > 2 && $user->getVerified() == false) {
+
+                $user->setVerified(true);
+                $job->setVerified(true);
+                $this->addFlash('notice', 'Oferta została pomyślnie edytowana oraz użytkownik został'
+                        . ' automatycznie zweryfikowany jako zaufany');
+
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($job);
+                $em->flush();
+
+                return $this->redirectToRoute('jobs_list');
+            }
+            if ($user->getVerified() == true) {
+
+                $job->setVerified(true);
+                $this->addFlash('notice', 'Oferta została pomyślnie edytowana');
+
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($job);
+                $em->flush();
+
+                return $this->redirectToRoute('jobs_list');
+            }
+
+            $em = $this->getDoctrine()->getManager();
+            $em->persist($job);
+            $em->flush();
+
+            $this->addFlash('notice', 'Oferta została pomyślnie edytowana i oczekuje na moderacje');
+            return $this->redirectToRoute('jobs_list');
+        }
+
+        return $this->render('Job/edit.html.twig', [
+                    'form' => $form->createView(),
+                    'job' => $job,
+        ]);
     }
 
     /**
-     * @Route("/oferta/{id}")
-     * @Template()
+     * @Route("/oferta/{id}", name="show_job", defaults={"id": false}, requirements={"id": "\d+"})
      */
-    public function showAction($id)
+    public function showAction(Job $job = null)
     {
-        return array(
-                // ...
-        );
+        $user = $this->getUser();
+
+        if (!$job) {
+            $jobs = $this->getDoctrine()
+                    ->getRepository('AppBundle:Job')
+                    ->findBy(array('user' => $user), array('publishedAt' => 'DESC'));
+
+            return $this->render('Job/show_jobs.html.twig', array(
+                        'jobs' => $jobs,
+            ));
+        }
+
+        $session = $this->get('session');
+        $lastSean = $session->get('lastFive', array());
+        array_unshift($lastSean, $job->getId());
+
+        if (count($lastSean) == 6) {
+            array_pop($lastSean);
+        }
+        var_dump($lastSean);
+        $session->set('lastFive', $lastSean);
+
+        return $this->render('Job/show.html.twig', [
+                    'job' => $job,
+        ]);
     }
 
 }
